Separate `CompareFilter` from `RangeFilter`
so that we can satisfy `oneOf` condition in the JSON schema.

Note that we need either https://github.com/YousefED/typescript-json-schema/issues/46
or https://github.com/YousefED/typescript-json-schema/issues/47 to
support combination of lt/lte and gt/gte in the schema.  (The logic works fine but those combo will fail to validate)

For now, let's support non-combo version for `CompareFilter`.