suggest to remove `schema` from `getting started`
Technically, a vega-lite specification can live without out the `$schema` property. Though this may not be recommended for many reasons (which are BTW not mentioned anywhere in the documentation), starting with `schema` in this context just breaks the flow. As a side node: the `vega` tutorials do not mention it either.