Enable initialization of scale-bound selections. (#5352)
We used to prevent scale-bound selections from being initialized  (users instead would have to explicitly define scale domains to initialize these type of selections). This is because selections used to be initialized via signals, which would populate the selection store on the second pulse and thus would not be ready in time for scale initialization. However, after switching selection initialization to happen directly in the store, this limitation no longer exists and scale-bound selections can be initialized directly rather than via the scale domain workaround.